+++ package/contents/config/config.qml
@@ -1,20 +1,14 @@
-import QtQuick 2.0
 import org.kde.plasma.configuration 2.0
 
 ConfigModel {
     ConfigCategory {
         name: "Behavior"
         icon: "preferences-desktop"
-        source: "config/BehaviorConfig.qml"
+        source: "config/BehaviorTab.qml"
     }
     ConfigCategory {
         name: "Appearance"
         icon: "preferences-desktop-display-color"
-        source: "config/AppearanceConfig.qml"
-    }
-    ConfigCategory {
-        name: "Advanced"
-        icon: "preferences-desktop-plasma"
-        source: "config/AdvancedConfig.qml"
+        source: "config/AppearanceTab.qml"
     }
 }