+++ package/contents/ui/common/GrowingTextField.qml
@@ -1,0 +1,15 @@
+import QtQuick 2.7
+import QtQuick.Controls 1.4
+
+TextField {
+    id: textField
+    enabled: root.enabled
+    implicitWidth: Math.max(30, hiddenTextInput.contentWidth + 16)
+    horizontalAlignment: TextInput.AlignHCenter
+
+    TextInput {
+        id: hiddenTextInput
+        visible: false
+        text: textField.text
+    }
+}