+++ scripts/install-applet.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+mkdir -p build && \
+cd build && \
+cmake .. && \
+make && \
+sudo make install
+cd ..