+++ scripts/install-dependencies-fedora.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+dnf install cmake extra-cmake-modules gcc-c++ qt5-qtbase-devel qt5-qtdeclarative-devel kf5-plasma-devel kf5-kglobalaccel-devel kf5-kxmlgui-devel