+++ scripts/install-dependencies-opensuse.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+sudo zypper in cmake extra-cmake-modules gcc-c++ libqt5-qtbase-devel libqt5-qtdeclarative-devel plasma-framework-devel kglobalaccel-devel kxmlgui-devel