+++ scripts/install-opensuse-deps.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-sudo zypper in cmake extra-cmake-modules gcc-c++ libqt5-qtbase-devel libqt5-qtdeclarative-devel libqt5-qtx11extras-devel plasma-framework-devel kglobalaccel-devel kxmlgui-devel